requirements.txt
paho-mqtt~=1.6.1
PyYAML~=6.0
cryptography~=43.0.0
aiohttp~=3.9.5
